chore: token-diet phase 1 — slim CLAUDE.md + split main.rs
CLAUDE.md 261 → 129 lines: moved P0a.5/P0a.6/Companion/GUI Export
historical detail to docs/architecture/runtime-overview.md.
Added new mandatory rule: single source file ≤ 800 lines.

Split mur-core/src/main.rs 2,457 → 152 lines:
- cli/mod.rs (412)        — Cli struct + top-level Commands enum
- cli/actions.rs (572)    — non-agent subcommand action enums
- cli/agent.rs (518)      — AgentAction + 10 sub-action enums
- dispatch.rs (687)       — Cli → cmd::* router (extracted from async_main)
- cmd/search.rs (150)     — cmd_search_unified handler

Pure code movement: 1,114 tests pass, clippy -D warnings clean.
Tests in main.rs preserved verbatim (3 ask-flag parser tests).

Reduces token cost per editing session by:
- ~2,000 tokens / turn from CLAUDE.md alone
- ~6,000-8,000 tokens / edit on the dispatcher (was 2,457-line read on every Edit)

Phase 2 (split outbox.rs / server.rs / cmd/agent.rs) tracked separately.
